chore: Refactor dowload-test-artifacts and test-renku for more generalization
The two actions are refactored to be more generic and not require
a values file, taking the reference to the S3 bucket holding
the test artifacts as input instead.

This is part of a larger work that will allow running the acceptance-tests
from the Github actions also for permanent deployments of Renku in addition
to the PR-deployments.